requirements.txt
functions-framework==3.2.1
google-cloud-storage==2.18.2
google-cloud-pubsub==2.25.2